Add support for .NNN file naming scheme
Note that this patch does not only add support for .NNN style
volumes but for any naming scheme (with extension .rar) using
incremental numbering. Be aware though that starting a volume
with a non-standard name that does not resolve to a 1
(e.g. .part.001.rar) will not work reliably. Especially volume
numbers starting with 0 (e.g. .000 or .part0.rar) might fail to
extract properly. This is however not a new limitation and has
been there since day one.

This patch should resolve issue #55

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>