Fix some corner cases in the directory cache logic
With the new directory cache there were a few corner cases missed
related to presentation of the files.
This patch specifically address files on the back-end fs having
conflicting names with files inside an archive and files with RAR
related extensions that was not really true RAR files or at least
could not be read as such.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>