Add static linking of UnRAR library
The UnRAR library built from the free C++ UnRAR source from RARLAB
is currently linked dynamically. Since the DLL extensions implemented
by rar2fs use headers outside that of the official API it might
result in compatibility issues. These issues can lead to reduced or
incorrect behavior or even crashes. Typically one such
incompatibility was reported in issue #62.
To avoid forcing a user to keep track of what exact version of the
source was used when building rar2fs and match that towards what
is actually installed on the system, static linking can be applied.
It is however not until very recently the UnRAR source tree provided
a static library build.

This patch will introduce a new build configuration option,
--enable-static-unrar. No actions are needed if static linking is
favored but if the option is not used a fall-back mechanism will
be applied. If a static library cannot be found in the same source
tree as from which rar2fs was built, dynamic linking (as done today)
will be performed instead. This is thus a backwards compatible change
for those that chose not to build a static library or simply have
not yet had chance to prepare one. If static linking of the UnRAR
library for some reason is not wanted the option --disable-static-unrar
can be used to completely override the new approach introduced by
this patch.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>