Fix file collection at archive mount
The function that collects the files to be part of an archive
mount is not able to properly see the difference between a real
volume and files simply named as such.
This patch adds an initial read of the RAR header to check for a
volume, and if it is, what format it has.
Also a problem with the password file parsing has been fixed
which caused additional problems for encrypted archives.

This should resolve issue #65.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>