Check consistency of new volumes
When new volumes are added the files might not be sent to the file system
in the order expected. This patch will make sure the critical files are
present and intact before contents will be made visible during header scan.
Also the keep_cache flag has been disabled since the effect of keeping the
cache between open calls during e.g. external library updates needs to be
further investigated. The performance loss due to this change should barely
be noticeable.

This patch should address most of the problems reported in issue #66.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>