Remove internal support for RAR in RAR
RAR files that include other RAR files is supported using
internal logic for up-to two levels. The problem with this is
that it requires a lot of extra and complex logic to resolve
the second level of information. Also it has limitations that
prevents it from working in all cases which makes it a bit
un-deterministic and requires more resources than might be
expected/accepted.
This patch will remove the internal support for multi level
archives and if such RAR files need to be resolved stacking
should be used instead. The option --seek-depth is completely
removed by this patch since it no longer serves a purpose.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>