Fix issue with SIGTERM not always working when daemonized
The function fuse_set_signal_handlers() is called before calling
fuse_daemonize(). This has some effect on signal processing that
on CYGWIN causes the SIGTERM signal to not be delivered when
issuing e.g. the 'kill' command on a daemonized rar2fs process.
This patch changes the order in which these two functions are
called.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>