Use proper definitions from libunrar headers
Many definitions are hidden and supposed to be used only internally
within the library itself, some of those having the same value(s)
as the public exported interface. Later version of the library will
properly export all public definitions instead of using magic constants.
This patch will prepare for those definitions by using them in the way
they were intended to be used. For backwards compatibility the
definitions are also added to the library extension if not already
defined. This patch also improves the presentation of volumes that are
missing or having corrupt key parts.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>