Fix parse problem for file header flags
When moving over to definitions from unrarsrc rather than using
internal hard-coded values, files started to show up as directories
instead of regular files as reported in issue #76. By accident this
worked before this change but the implementation was actually broken
even before that.
This patch fixes the incorrect header parser in a backwards compatible
way to work seamlessly also for earlier versions of unrarsrc.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>