Workaround for misbehaving pread(2) on Cygwin
At EOF pread(2) should return 0 but on Cygwin that is not the case and
instead some very high number is observed like 1628127168.

From the pread(2) man page:
  "On success, the number of bytes read or written is returned (zero indicates
  that nothing was written, in the case of pwrite(), or end of file, in the
  case of pread()), or -1 on error, in which case errno is set to indicate the
  error."

Let's assume that if what is returned from pread(2) is greater than the
requested read size, EOF has been reached.
Without this workaround at least dokan-fuse gets very confused and is
caught in an infinite loop!

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>