Show both local- and RAR-files in shared directory
This patch resolves issue #114

When listing a directory which exist both in an archive as well
as on the local file system only the local contents are shown.
Originally this was by design to simplify the directory cache logic.
After some investigation it was discovered to only impose an
unnecessary limitation as described in issue #114. Instead of hiding
the RAR contents instead scan also the archives and merge the
contents sharing the same folder. This patch does not affect the
behavior of the file system in any other way except from what is
described above.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>