Clear unexpected flags in open for WinFSP/cygfuse
When using WinFSP/cygfuse the O_CREAT and O_EXCL flags sometimes
seems to be added to fi->flags in the open callback function.
This is not according to the legacy FUSE API implementation for
which these flags bits are never set. As a workaround, make sure
to clear flag bits that are not expected in calls to open.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>