Rework seek length detection
Due to that at least two volume files must be searched for meta
data the seek length was automatically adjusted and incremented
by one. This however results in an unnecessary number of files
being opened. The seek length is now adjusted only in the case
that it has been set to 1, in all other cases the value is left
as is. Thus setting the seek length to 1 or 2 effectively means
the same amount of files are inspected.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>