Detect source tree updates
Commit 1553f5c661bc1819e171577b079e084a39a221b5 introduced the
directory cache. The increased performance came at the cost of
lost synchronization with the source tree. New archives added
to a source directory does not show up unless the cache is
invalidated. This is normally not an issue since changes should
be made through the mount point rather than the source tree,
thus avoiding the synchronization problem since rar2fs is notified
about the change. It is however not always feasible to allow
writes (-o rw) to the mount point.
Add the directory modification time (st_mtim) to each cache entry.
At cache hit compare current directory st_mtim with what is saved
in the cache. A possible mismatch will invalidate the cache entry.
Adding these actions will reduce the possibility of a mismatch
between the source tree and the mount point significantly, but
unfortunately it does not eliminate it completely. This because the
directory modification time is not updated if e.g. an existing file
is replaced/overwritten rather than removed and then added again.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>