Fix a potential stack allocation overflow
When the RAR file contents is resolved a stack allocator is
used for temporary storage of complete path names.
For RAR files with a lot of resident files and directories the
allocation might overflow available stack memory leading to
unpredictable behavior and a potential program crash.
Replace the stack allocation with a traditional heap allocation
instead.

This problem was first seen and reported in issue #118 but has
been a potential issue for a very long time.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>