Change FUSE library search order
The order in which the different supported FUSE implementations
are searched for is a bit off. This has caused some issues on e.g.
OSX for which installations of OSXFUSE and the old MacFUSE
compatibility layer co-existed. An application today should never
link towards the MacFUSE compatibility layer. It exists only to
support legacy MacFUSE file systems binaries that might still be
floating around. This problem was reported in issue #119.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>