Fix all uses of dirname(3)
Commit 1d6bf47979d9d15a0d4cfc48ddaee5a8716e88a9 fixed a few places
in which calls to dirname(3) was wrongly assuming that its input
argument was modified. The fix was to instead use the returned string
from the function. This is how ever not enough since those implementations
many times also seems to rely on a common output buffer in which
consecutive calls to dirname(3) might alter a previous returned string
which thus cannot be trusted to remain the same for any given time or
scope. This means that to use dirname(3) safely code has to assume it
is both non thread safe and input modifying.
Avoid this pitfall by implementing a local dirname-like function with
known semantics across all platforms.

This problem was discovered during trouble shooting of issue #119.
The unsafe and incorrect use of dirname(3) would in theory not only
affect Mac OS X 10.15 but a lot of versions before that, if not all.
The actual problem observed was however introduced very recently in
commit 06b9d54545d13c224967fc64fe7bce5a51c953a9 which has not yet
been officially delivered. All other potential problems with dirname(3)
addressed by this patch are less probable to manifest themselves.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>