Fix cache invalidation for child entries
There is a flaw in that a cache invalidation does not also invalidate
child entries. A typical problem with that is if e.g. a directory is
renamed, the child entries will be left in limbo, basically resulting
in a memory leak. Another issue is that if a RAR archive is replaced by
a new version, old stale cache entries might resolve to historic
information due to a false cache hit.
This patch addresses part one of the problem reported in issue #124.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>