Add optional cache warmup
This is an enhancement originating from issue #131.
There are use-cases in which it makes sense to trigger a background
warmup of the internal directory- and file caches. The warmup does
not conflict with normal file system operations. The cache warmup is
optional and by default disabled. To enable the background cache
warmup use the new mount option `-owarmup` or `-owarmup=<n>`.
The latter syntax can be used to tweak the number of background workers.
The default number of workers is 5.

The rar2fs man-page has been updated accordingly.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>