Improve speed when listing archives with many directories
The population of the directory cache is "lazy", which basically means
it is not populated until a directory is queried by e.g. a readdir() call.
The effect of this is that the entire archive is scanned multiple times,
ones for each directory query. Become more aggressive in the population of
the directory cache. Basically making sure each directory in an archive is
added to the cache at first scan, even if it was not specifically the
subject of the scan itself.
This patch addresses part two of the problem reported in issue #124.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>