Change FUSE library search order
The library search order is currently making OS X based platforms picking
up the wrong library if the MacFUSE compatibility layer is installed.
This is due to that naming of the library for systems not using 64-bit
I-nodes matches exactly that of the legacy Linux FUSE library. Change order
in which the libraries are searched and first check for libraries that will
not conflict with legacy. This patch also adds support for a missing 64-bit
variant of the OSXFUSE libraries (libosxfuse_i64).
The problem with the library search other was first reported in issue #138.

Note that not installing the MacFUSE compatibilty layer as part of the
OSXFUSE package requires a --with-fuse=/usr/local/include/osxfuse directive
to `configure` or it will fail to locate the required FUSE header files.
The MacFUSE compatibiliy layer is known to be deprecated and it is advised
not to be used.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>