Require password for encrypted archive mounts
Both archive and folder mounts require a password only in the case that
the archive header is encrypted. When mounting a single archive this
behavior is not very useful. For single archive mounts require a password
also if data is encrypted.

This patch does not change the behavior when mounting folders for which a
password is never required to view the file list when the header is not
encrypted. A flaw was however fixed that could lead to garbage data being
extracted rather than returning an error if wrong password was provided.
Note however that in the way the RAR4 format works, extraction of
encrypted archives with CRC errors will still be detected the same way as
if the wrong password was provided. For that reason the user feedback has
also changed to become a pure input/output error to be consistent across
both RAR4 and RAR5 formats.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>