Make sure a bad password is detected for RAR4
This was another difference between RAR4 and RAR5 that had to be
addressed. If a bad password is given it was enough to try to open a
RAR5 archive with encrypted headers. This is not the case for RAR4.
For simplicity treat both formats the same and always scan the
header for meta-data.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>