Reorder process startup sequence
For archive mounts move the gathering of volume files is done after input
paths have been resolved completely rather than being part of the function
itself. This change was necessary to make sure the correct absolute paths
were available when scanning archive headers and to allow for the new
.rarconfig settings to be available also for archive mounts. Previously
this was limited to folder mounts only.

This patch also includes some other small changes and improvements
including the removal of the verbose flag which had little to no use.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>