Fix a few bugs with different severity
- An error case in the implementation of open() could result in that a
undefined pointer is dereferenced at memory allocation failure. This
will result in process termination due to a segmentation violation.
This is a major bug yet very unlikely to happen.

- Avoid a superfluous extra lookup of "/" at readdir() operations. This
will reduce the number of needed archive scans for the root folder until
the directory cache is populated.
This is a minor and harmless bug.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>