Fix race condition at directory cache invalidation
Commit 33af7aab6b224f9a42d57831769b22fe93f2678e did not completely
solve the problem(s) reported in issue #151.
After running valgrind/helgrind a race condition was spotted with
the following signature:

==29979== Possible data race during write of size 8 at 0x62E2F20 by thread #7
==29979== Locks held: none
==29979==    at 0x411D6B: hashtable_entry_delete_hash (string3.h:52)
==29979==    by 0x412180: hashtable_entry_delete_subkeys (hashtable.c:250)
==29979==    by 0x4126EE: dircache_get (dircache.c:195)
==29979==    by 0x41BC63: syncdir (rar2fs.c:3082)
==29979==    by 0x41BF12: rar2_getattr (rar2fs.c:3203)
==29979==    by 0x4E4534F: lookup_path (fuse.c:2472)

The problem is caused by a missing rwlock when detecting stale cache
entries and when such entries were invalidated.
This patch also adds a few missing locks spotted by pure inspection
of the code and which were not part of the use-case covered by the
valgrind/helgrind test run.

Resolves-issue: #151
Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>