Change approach for cache invalidation
Change to a recursive (faster) approach when invalidating directory
cache child entries. The recursion depth is set to a limit that if
reached will back off to the previous linear bucket search (slower).
Using this approach should gain some performance for the most use-cases
in which folders are modified on a regular basis and will only inflict
some minor performance penalty in more extreme situations.

Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>