Fix crash in collect files()
When collect_files() tries to identify first volume file there is a
missing check on a pointer to be valid or not. There are two identified
scenarios in which this can happen.

1) Failing to allocate memory that in current design could still produce
a pointer that is not NULL but still not valid. This is due to how C++
'operator new' by default works since the time exceptions were introduced.
In this case simply checking for NULL would not be enough.

2) Probably a lot more likely but still must be considered an extremely
rare case is that the function involved returns ERAR_EOPEN without
first having allocated this memory in the first place. This can only
happen in a very specific branch in the unrar library that has been
overlooked or it was introduced in a later version of it. It covers a
very old RAR compression format/version that had a flaw in the way
length of each file was calculated. It could result in that an entire
file fitted in one volume but still indicated in the header that it
continued in the next. Such archives must not be allowed to be opened
and hence an error is thrown. A simple test for NULL would cure that
since pointer is initialized to 0 and if no allocation is performed it
will remain 0 across the function call.

Thus to solve both these scenarios the solution is to:
  - check for NULL where relevant
  - call operator new with the (std:nothrow) argument

This patch also introduce some stability changes to collect_files()
and should avoid some other potential problems for obfuscated and
non-standard archive names.

Resolves-issue: #167
Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>