Make sure reader thread is terminated properly
Under some conditions the reader thread responsible for carrying data
from the child process to the I/O buffer may fail to terminate at close
and file system thread gets blocked waiting for join to complete.
Terminate the reader thread using a control message rather then using
the brute force call to pthread_cancel().

Additional updates made in this patch to increase robustness include:
  - perform dry-run extraction outside of child process and before it
    is spawned

Note that the majority of the changes done as part of this patch only
affects extraction of compressed (-m1 and above) and/or encrypted archives.

Resolves-issue: #165
Signed-off-by: Hans Beckerus <hans.beckerus at gmail.com>